backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.10.0
python-dotenv==1.0.1
pandas==2.2.3
httpx==0.27.2
pydantic==2.9.2
python-multipart==0.0.12
groq==0.11.0
yfinance==0.2.43
